refactor: move method requiring cryptography to the appropriate file
